+++ features/notification/apis/dto/notification.ts
@@ -1,16 +1,6 @@
 import { Response } from '@/apis';
 
-export interface NotificationElementProps {
-  type: 'CHEER' | 'FOLLOW' | 'FRIEND';
-  notificationId: number;
-  nickname: string;
-  profileImageUrl?: string;
-  memberId?: number;
-  content?: string;
-  createdAt: string;
-  recordCreatedAt?: string;
-  hasRead: boolean;
-}
+import { NotificationElementProps } from '../../types';
 
 export type NotificationResponse = Response<{
   notifications: NotificationElementProps[];
