+++ features/notification/index.ts
@@ -1,2 +1,3 @@
 export * from './apis';
 export * from './components';
+export * from './types';
