--- features/notification/type.ts
@@ -1,10 +1,0 @@
-export interface NotificationElementProps {
-  id: number;
-  variant: 'follow' | 'cheer';
-  userName: string;
-  time: string;
-  isFollowing?: boolean;
-  description?: string;
-  recordDate?: string;
-  isClicked: boolean;
-}
